requirements.txt
asgiref==3.3.4
Django==3.2.3
django-tempus-dominus==5.1.2.13
django-widget-tweaks==1.4.8
pytz==2021.1
sqlparse==0.4.1
